+++ src/Navigation/index.tsx
@@ -1,9 +1,9 @@
 import { enableScreens } from 'react-native-screens';
-import { DrawerContent } from './DrawerNav';
+import { DrawerContent } from './drawer-nav';
 import { createDrawerNavigator } from '@react-navigation/drawer';
 import { RootDrawerParamList } from '../types';
 import React from 'react';
-import RootTabNavigator from './RootTabNavigator';
+import RootTabNavigator from './root-tab-navigator';
 import Logo from '../assets/logo';
 
 //https://docs.expo.dev/versions/latest/sdk/screens/ -Kai
